refactor(orb): extract protocol constants + greeting pools to orb/ tree (BOOTSTRAP-ORB-MOVE)
Phase 2 (move-only) partial — extracts pure constants + multilingual static
data from routes/orb-live.ts to the new orb/ module tree. Zero behaviour
change: every value is preserved byte-for-byte, imports re-bind identifiers
back into local scope at call sites.

Extracted to orb/upstream/constants.ts:
  - VAD_SILENCE_DURATION_MS_DEFAULT
  - POST_TURN_COOLDOWN_MS
  - SILENCE_KEEPALIVE_INTERVAL_MS, SILENCE_IDLE_THRESHOLD_MS,
    SILENCE_PCM_BYTES, SILENCE_AUDIO_B64
  - GREETING_RESPONSE_TIMEOUT_MS, TURN_RESPONSE_TIMEOUT_MS
  - FORWARDING_ACK_TIMEOUT_MS
  - MAX_CONSECUTIVE_MODEL_TURNS, MAX_CONSECUTIVE_TOOL_CALLS
  - connectionIssueMessages (per-language user-facing string pool)

Extracted to orb/instruction/greeting-pools.ts:
  - SHORT_GAP_GREETING_PHRASES (8-language pool, ~90 lines)
  - pickShortGapGreetings (randomizer)

Result: orb-live.ts shrinks by ~180 lines. First real step of the Phase 2
module tree forecast in the atomic-riding-badger plan. Follows Phase 6
scaffolding (#757) which established the orb/ directory; subsequent moves
(watchdog + keepalive + reconnect, system-instruction builders, tool
execution) continue shrinking the monolith file.
